requirements.txt
python-dotenv==1.0.1
requests-toolbelt==1.0.0
aiohttp==3.9.5
python-magic==0.4.27
aiofiles==23.2.1